Ensure that methods are invoked after the resource manager was created
With the default order, new method invocations are added after the
object was created. If one of those invocations uses our resource
manager, then it might happen that this call is added before the object
has been created, leading to a NullPointerException.

To avoid this, we use a custom order where we first check, where this
resource manager is created and then determine the location in the
constructor, after which new methods can be added.