Consider the correct font when calculating the palette entry bounds
The width and height of the palette entries (entry/components) is
calculated using the font of the palette composite.
Given that the user can set larger fonts for the entries, this may then
end up in a situation where the text is larger than the figure, meaning
the edges of the text are cut off.
This is fixed by using the correct font, in order to calculate the
correct bounds of the figure using the Draw2D FigureUtilities class.