Use JDK classes instead of Guava to create immutable collections
Java itself is powerful enough to express immutability. There is no need
to keep using Guava for this. This change removes all usages of
ImmutableList, ImmutableSet and ImmutableMap and replaces them with
List, Set, Map and the Collections utility class.

Example:
ImmutableList.copyOf(...) -> List.copyOf(...)
ImmutableMap.of(...) -> Map.of(...)
ImmutableSet.of() -> Collections.emptySet()