[GTK] Cleanup SWT snapshot mechanism on Linux systems #609
This change consists of the following part:
- The unused methods have been removed from the native C code. They were
still called by the 32bit implementation. But given that this mode is no
longer supported by SWT, there is no point in keeping it.

- The snapshot mechanism exclusively uses the native implementation (how
it used to be). This should hopefully avoid issues where a snapshot is
taken before the widget has been rendered.

- We no longer attempt to capture the decorations. Those are part of the
window manager and not the shell itself. On X11, it is still possible
(yet unreliable) but on Wayland it seems to be prohibited to access
anything outside the widget.
Doing so also causes several issues where the figures are cut off or
offset by the insets of the window.

Should resolve #609