Try to minimize "Source folder not found" error during Windows tests
Some of our tests occasionally fail due to a missing source folder in
the test project. What's exactly causing this is not immediately
obvious, but given that it only happens on Windows, the most likely
candidate is its file management system.

If this case occurs, we now simply create a new source folder from
scratch (while logging a warning), instead of simply throwing an
exception.