Add required modules to module-info.java file when creating new template
In case a project contains the module-info.java, the user has to
manually add e.g. the "java.desktop" module, if they want to use Swing.
If not, WindowBuilder will be unable to parse the source code and thus
also unable to open the design page.

Given that in the project wizard, the checkmark to create this file is
set by default, so it's very common to create and just forget about this
file, when setting up a workspace.

Figuring out which modules to add is quite tedious and not immediately
obvious for those not experienced with this feature, meaning we should
attempt to automate this when possible.

Resolves #472, #473, #574, #592, #613