Restore delayed event execution for EditEventManager
In an earlier commit, the delayed execution of events was moved to the
dispatch methods, instead of being implemented in the listeners
directly. Because those methods are overwritten by the EditEventManager,
this mechanism was no longer applied here.

This causes an exception when e.g. trying to drag an SWT Group into the
design canvas for the first time.

Amends 9f074b858f5b5425864ecf669b6a100c417033c3