[GEF] Resolve DragTracker ambiguity in EditPart
In order to get the DragTracker of a given EditPart, one can either call
getDrackTracker (GEF) or getDrackTrackerTool(WB). This change refactors
the our code so that only the GEF method remains.
Note that this change requires a noticeable cleanup in our Tool
implementation, in order to make sure that the methods defined by the
Tool interface don't call the implementations of the parent class.