[GEF] Remove local EditPolicy-List variable from EditPart
This removes the local fields that store the policies (and keys)
installed for a given edit part. All invocations are forwarded to the
implementation of the parent class.

Note that those methods work on a GEF EditPolicy. To avoid any
incompatibilities, it was required to also remove the local "host" field
of our EditPolicy and rely on the parent implementation here as well.