Repair combo-box in layout preference page #640
Selecting any element in the default layout manager combo box always
reverts back to the implicit layout. But this should only happen, if the
element is unchecked in the table below.

The input of the combo is adapted whenever the checked state of the
table changes, to ensure that the user is only allowed to set those
layouts as default, which are checked.

This change also converts the raw SWT widgets into their matching JFace
viewers, to allow us to work directly on the LayoutDescription objects,
rather than on e.g. Strings.