Remove XWT integration for upcoming WindowBuilder release #645
The project is no longer actively maintained and with the Eclipse
2023-12 release, it is also no longer possible to resolve our workspace.

To avoid the disproportional amount of work that goes with keeping that
project alive as well, it seems to be the safest solution to simply
remove it from our codebase.

Note that this is also the reason why neither XWT nor our XWT extension
have been part of the 2023-12 release train.