Remove calls to Guavas Lists/Iterables methods #270
In almost all cases, the call of Lists.newArrayList() can be simply
replaced with new ArrayList<>().

List.transform() or Iterables.filter() can be replaced by the Java
Stream API, which is also encouraged by their JavaDoc.

Occasionally, we use List.of() to create lists from arrays. But care has
to be taken here, because those lists are read-only. If values are
written to them afterwards, the constructed list is passted as an
argument to an ArrayList.

The MapMaker class which is used to create maps with weak keys/values
are replaced by the ReferenceMap.

The Multimap class is replaced by the MultiValueMap.