Remove usage of Guava Collections in WindowBuilder Core #270
In almost all cases, the call of Lists.newArrayList() can be simply
replaced with new ArrayList<>().

Occasionally, we use Arrays.asList() to create lists from arrays. But
care has to be taken here, because those lists are read-only. If values
are written to them afterwards, the constructed list is passed as an
argument to an ArrayList.
Note that we use Arrays.asList() instead of List.of(), as latter doesn't
allow null elements.

The MapMaker class which is used to create maps with weak keys/values
are replaced by the ReferenceMap.

The Multimap class is replaced by the MultiValueMap.