Create new, dedicated plug-in to generate Java code from XSD
The Java code is generated automatically via the jaxb2-maven-plugin and
therefore not checked in. With the latest m2e version, this Mojo should
be executed as part of the incremental build.
By using those POJOs, we are able to process the XML extension
contribution instead of Digester3.

Note that processing the generated code is done via the JAXB API.
Meaning that this approach adds a dependency to jakarta.xml.bind (and
jakarta.activation). While the project depends on the latest version, we
are compatible with all versions starting from 3.0.0.