Process the (standard bean) properties of a component via JAXB
The "parameter" and "parameter-list" elements of the property editor and
configurable type have been moved to a common sub-type so that they can
both be handled by the same method.

The "PropertyConfigurationElements" group has been converted to a
complex type, extended by both the "PropertyConfiguration" and
"MethodSingleProperty" type, in order to be handled by the same method.

The rules have been adapted to accept both Digester3 and JAXB
instructions. This is because they are still used to process other
properties, which still use Digester3, which would then inflate the
amount of change that needs to be done per commit.

See #638