Use absolute coordinates for start and end location while dragging
When dragging the cursor (e.g. to resize a figure), the corresponding
tool would use the absolute coordinate of the cursor as start point, and
the relative coordinate as end point when determining whether dragging
should be enabled.

This fails once the user scrolled slightly to the right/down. To fix
this, simply replace the call to getCurrentInput().getMouseLocation()
with getLocation(), which returns the ABSOLUTE mouse location.

Resolves #697