[Win32] Repair selection in WindowBuilder palette
The deselection of the palette entries is not registered on Windows.
While the entry can be selected, it remains selected even if other
entries are chosen.

This is because the setCapture() method in Figure was removed and
therefore those method calls in the PaletteComposite now reference the
Control method instead. The reason this doesn't affect Linux is because
here this is a no-op method.

Amends 360a0a678fb52b7a8e33b5fbb97de7967e6cc069