[Draw2D] Paint property values using Graphics instead of GC object
In preparation of converting the property table into a proper GEF
viewer, the dependent classes need to be adjusted to use the Draw2D API.

This change only wraps the GC object in a Graphics object, which is is
then passed as an argument to the property editors. Internally, this
Graphics object then calls the corresponding GC methods.

In the future, this Graphics object is supplied by the viewer when
painting the Draw2D figures and no longer needs to be constructed by us.