Refactor SWT baseline plugin to fix SpotBugs warnings
This resolves the following warnings:
- SING_SINGLETON_HAS_NONPRIVATE_CONSTRUCTOR
- SING_SINGLETON_GETTER_NOT_SYNCHRONIZED

One warning is created in the Activator. Given that this class is not
used, we can simply remove it.
For the Baseline class, this is a little bit more difficult, because a
constructor is needed for the platform-specific subclasses. Here a
combination of a base interface and the "sealed" keyword is used.