[SimRel] Prepare workspace for 2024-06 release
This change updates all features to 1.16.0, updates the target platform
to use the 2024-06 repository and handles the changes detected in the
XSD plugin.

To handle the timestamp in the generated JAXB files, a PR has already
been created. In the meantime, we simply replace it using a small Ant
script.