Use AbstractUIPlugin instead of ResourceManager to load ImageDescriptor
This removes the internal usage of
ResourceManager.getPluginImageDescriptor() in favor of
AbstractUIPlugin.imageDescriptorFromPlugin().

Note that the ResourceManager class is using a special
PluginResourceProvider interface for loading bundle images at design
time, because the widget class is not parsed within an OSGi environment
and therefore Platform.getBundle() always returns null.
To replicate this behavior, the method invocation is hijacked and
replaced by a call to SwtInvocationEvaluatorInterceptor.getEntry() (the
same method that is used by the ResourceManager).

See #486