Add ability to contribute OSGi bundles as additional libraries
The layoutManagers and toolkits extension points have been adapted to
not only support plain Java jars, but also proper OSGi bundles. To
properly handle both types, the LibraryInfo class has been split into a
JarLibraryInfo and a BundleLibraryInfo variant.

For bundles, we also compare the version specified in the bundle
manifest with the bundle in the local classpath. If the version in the
workspace is older than what's used by WindowBuilder, the jar is
replaced.