Consume JGoodies bundles from Orbit instead of using embedded jars
This removes the "org.eclipse.wb.swing.FormLayout.lib", which was used
to wrap the plain jars "com.jgoodies-common" and "com.jgoodies-forms".
Instead the bundles from Eclipse Orbit are used.

Note that this also updates JGoodies Common to 1.8.1 and JGoodies Forms
to 1.9.0, which are the last releases which are open source.