Fix incompatibilities introduced with 2024-06 M2
The method Shell.adjustTrim() has been replaced by a new method (with
same name), which now expects two integers as as arguments. For the sake
of backwards compatibility, we check the version of SWT and then call
either one or the other method.

Furthermore, the memory leak in ClassUtils has been fixed, so we no
longer need to handle this case in our test suite.