Evaluation of array should also consider assignments after creation.
The current implementation ignores all changes to an array that are done
after it has been initialized, which has the following two effects.

- When the array is used for display purposes, the composite that is
shown in the editor no longer matches what is visible at runtime.

- When non-primitive types are used, this might lead to
NullPointerExceptions while handling the execution flow. In the worst
case, an internal error is thrown and the editor crashes halfway
through.

To fix this, we need to keep track of all such assignments and evaluate
them, when calculating the value of the array.

https://github.com/eclipse-windowbuilder/windowbuilder/issues/747