Integrate presentation buttons of property table into GEF layout.
Instead of placing the buttons directly on the canvas, they are now
wrapped inside a Draw2D figure.

Via the paint() and erase() method, we control whether the buttons
should be shown or hidden. The bounds of the buttons are described by
the bounds of the figure.

To make this possible, the return value of the show() method has been
moved to a seprate getSize() method. This is necessary to calculate the
size of the figure without creating (and showing) the SWT widget.