Make paining of expanded lines more robust.
When a property is expanded, we draw a line encapsulating all
sub-properties. We calculate this line by multiplying the column height
with the number of children. The problem here is that this approach no
longer works when spacing is taken into consideration.

To fix this, get the edit parts for each property and calculate the
length by using the bounds of figures enclosed by the first and last
edit part.