Fix: Only disable resizability of frame when taking a screenshot.
We explicitly cann frame.setResizable(false), in order to allow taking a
screenshot for composites that are larger than the screen resolution.
This has the unpleasant side effect that it's not possible to maximize
the composite when opening the preview mode.

This property is now set before the screenshots are created and unset
once everything is finished.

Fixes https://github.com/eclipse-windowbuilder/windowbuilder/issues/751

Amends 8c91b9b544942f659e93d824d9a22aaf3cabb5df