Fix:Layout preference page doesn't keep default layout
The input of the combo viewer needs to be set before the databinding is
created. Otherwise the initial selection can't be restored. When
unchecking an element in the table, the old selection of the combo
viewer needs read before the new input is set, otherwise this
information is lost.

Fixes https://github.com/eclipse-windowbuilder/windowbuilder/issues/785