Always show scrollbar in property table to avoid obscured buttons
This effectively disables the scrollbar overlay so avoid weird
situations where buttons can't be clicked because they are obscured by
the srollbar itself.

Furthermore updates to the latest GEF build in order to fix other issues
where the scrollbar bounds are not considered when updating the bounds
of the root figure.

Fixes https://github.com/eclipse-windowbuilder/windowbuilder/issues/795