Fix partial initialization of ImplicitLayoutCreationSupport
When a new instance of ImplicitLayoutCreationSupport is created, one
must make sure that setJavaInfo is called. This is normally done when
calling JavaInfoUtils.createJavaInfo().

Failing to do so results in a AssertionFailedException, when an object
is removed from the container again. To reproduce, simply create a
JFrame, add a JPanel and then immediately remove it again.

Amends 280f8aa3282dce10f92e7d2cbe7077071aa11836
Fixes https://github.com/eclipse-windowbuilder/windowbuilder/issues/813