[GEF] Remove duplicate activate/deactivate methods in edit part
With the exception of calling the fireActivated() and fireDeactivated()
methods, the active() and deactivate() method of the GEF and
WindowBuilder edit part are functionally identical. Given that we don't
register any of those listeners to the edit parts, they become
interchangeable.

The registerEditPart() and unregisterEditPart() methods of the viewer
are replaced with the register() and unregister() methods within the
edit part itself. In addition to registering the model, this method now
also registers visuals and the accessibility edit part. Neither of which
we currently use.