[GEF] Harmonize selection methods in IEditPartViewer base interface
Both the IEditPartViewer and the EditPartViewer interfaces specify a
selection API. One is using the GEF edit parts, the other the
WindowBuilder edit parts. Because those methods have the same name but
not the same signature, one doesn't overwrite the other and it is
therefore ambiguous which one is actually called at which point.

To solve this, the API has been removed from IEditPartViewer and all
invocations have been adapted to support the GEF edit parts.

See https://github.com/eclipse-windowbuilder/windowbuilder/pull/829