[Draw2D] Harmonize Parent/Children accessors in Figure with base class
This makes it so that the add(), remove() and setParent() methods of our
Figure class match the ones defined by the GEF figure. Doing so avoids
any ambiguity regarding which list a child is added to.

One noticeable difference between the GEF figures and our figures is
that we don't use a layout manager and instead treat the constraint as
bounds of the figure-to-add. This mechanism is only used for the
OutlineImageFigure. In order to adapt those figures to the new behavior,
they need to explicitly told to use the XYLayout.