Create 96 and 192 DPI variants of the alignment icons.
This converts all of the alignment icons from gif to png and also adds a
192DPI variant. In order to encourage reusability, those images are kept
in the common base plug-ins, rather than each plug-in keeping its own
copy of the image.

In order to keep changes manageable, we only provide an interface for
accessing those icons, but don't actually use them. Future changes need
to make sure that the old icons are removed and their replacements
accessed via the new (internal) API.

This contribution also contains the original SVG files easily allow the
creation of higher resolutions of those icons, if necessary.