Fix: HTML Tooltip not show on Linux
A "location changing" event is fired while the tooltip is constructed on
Linux (but not on Windows). Because the listener for this event disposes
the tooltip shell, the browser is closed before it is completed,
resulting in the tooltip never being shown.

The purpose of this listener is to close the tooltip when a hyperlink
has been clicked and to open the URL in the normal web browser. It is
therefore sufficient to only add this listener AFTER the tooltip has
been created and made visible, therefore ignoring all intermediate
events.

Fixes https://github.com/eclipse-windowbuilder/windowbuilder/issues/842