Remove old alignment icons for Spring-/From-/Gridlayout
This removes the alignment icons that are contributed by said icons, as
well as their the accessors, where appropriate. Those icons are no
longer used in favor of the ones contributed by the CoreImages class.

More specifically, those are the layouts that are defined by the
AbsoluteComplexSelectionEditPolicy class, which no longer implements the
IActionImageProvider interface