[GTK] Use primitive types in native method invocation.
The native bindings where created with both 32 and 64 bit in mind. As a
result, the native methods are invoked using Integer and Long objects,
rather than primitive types, causing a lot of boxing and unboxing within
the C layer. With this change, we exclusively use the long type to
handle memory addresses, making 64bit the "default". Any architectures
with shorter memory addresses have to cast their values to e.g. an int.

Support for older architecturs (if ever needed) is done on the C layer
via an explicit type cast (utilizing the pre-processor) when calling the
proper GTK methods. While this introduces a small overhead for
non-64-bit architectures it is very likely compensated by no longer
having to work with the Java objects.