Fix: NPE when creating snippet without selecting a project
The IJavaProject instance returned by the Design wizard is based on the
project selected in the Project Explorer, rather than the project in the
dialog. If no project is selected, this then leads to a
NullPointerException. Instead, the project referenced in the main page
is used.

To avoid similar issues in the future, the WizardUtils class is removed
(as it isn't used anywhere else) and the tests have been adapted to also
consider this case.

Note that the Wizard tests have been migrated to SWTBot, as the
UIContext class (which has now been marked as deprecated) may get stuck
indefinitely, if an exception is thrown. Reason being that both the IDE
and the test suite compete over the UI thread, which can easily result
in a deadlock. As a result, the test suite has been split up into two
executions; One which is executed in the UI thread and one which isn't.

Amends 4af75eafcd4a7d97ba32ae96160e68709260541a
Fixes https://github.com/eclipse-windowbuilder/windowbuilder/issues/880