Explicitly set required permissions for "Publish Test Results" action
The workflow action no longer creates a comment containing the test
results on a given pull requests. Based on the log, it seems to be
unable to find the PR...

This is an attempt to solve this issue by giving write-access to pull
requests, as listed by the minimal workflow job permissions.

See:
https://github.com/EnricoMi/publish-unit-test-result-action/blob/master/README.md#permissions