[GTK] Harmonize signature of native getWidgetBounds() method
This renames the getWidgetBounds() to _gtk_widget_get_allocation() and
adapts the signature to expects a GtkAllocation object.

On the C layer, this Java object is converted to a C object using the
JGtkAllocation struct. This struct keeps track of all fields of the Java
class and used to exchange the values with the Java layer.

See https://github.com/eclipse-windowbuilder/windowbuilder/issues/871