Make scroll-bar invisible for horizontal/vertical header
GEF automatically adds the H_SCROLL/V_SCROLL styles when creating a new
FigureCanvas instance, leading to scroll-bars being shown in the headers
and thus obscuring the entries.

Set the visibility of the scroll-bars to never be shown, to get around
this problem. Note that those headers scroll together with the main
viewer and the scroll-bars are therefore not needed.

Relates to:
https://github.com/eclipse-windowbuilder/windowbuilder/discussions/894