Drop support for JSR296 (Swing Application Framework)
This specification request has been withdrawn in 2011, the latest Maven
jars date back to 2010 and the official website went offline a long time
ago.

In short, the project has been dead for over a decade and all major have
migrated to a modern framework a long time ago. It's only sensible for
us to follow suite.

Resolves
https://github.com/eclipse-windowbuilder/windowbuilder/issues/887