Cleanup and decouple SWTBot from other JUnit tests
This removes the dependency from the AbstractSWTBotTest class to the
AbstractJavaProjectTest class and therefore makes the SWTBot tests
independent from the "old" tests.

This dependency was used to perform I/O operations, which is problematic
due to its dependency on the UI thread. Instead, helper methods have
been added to our custom SWTBot instance, which take over this task.