Deploy test bundle to update site for baseline checks
The baseline check for the test bundle only works if the update site
contains a version to compare it to. Without it, any required version
bumps are missed, as they don't cause errors.

All packages exported by this plugin are marked as internal, to allow
the addition, modification and removal of tests without breaking "API".