[eRCP] Remove obsolete RectangleSupport class
Primarily used when converting the widget bounds to a Draw2D rectangle.
For several subclasses of the WidgetInfo, the getWidget() method has
been overridden, to expose the underlying widget and to avoid having to
call e.g. "getBounds()" via reflection.

As a result, this now means that the plugin containing the Nebula
definitions now explicitly depends on the Nebula widgets.

Contributes to
https://github.com/eclipse-windowbuilder/windowbuilder/issues/942