Add typified getter for the Object contained inside a JavaInfo
This makes it possible to have the method return the actual type that is
contained by a JavaInfo (e.g. a Button for a ButtonInfo), which then
makes it possible to access the relevant fields and methods directly,
without having to rely on reflection.

By using the public API, the code should not only become much more
readable, it should also highlight incompatibilities when fields and/or
methods are no longer accessible.

Note that we can't do the type-cast directly in getObject(), as it is
not guaranteed for a ComponentInfo, in case the component differs from
its visual presentation, to have a matching type, which would then lead
to a ClassCastException during runtime.