Remove "final" modifier of JavaInfo.getObject() and override in subtypes
This makes it possible to have the method return the actual type that is
contained by a JavaInfo (e.g. a Button for a ButtonInfo), which then
makes it possible to access the relevant fields and methods directly,
without having to rely on reflection.

By using the public API, the code should not only become much more
readable, it should also highlight incompatibilities when fields and/or
methods are no longer accessible.