Raise PDE dependency to require at least the 2024-06 release
One of PDE bundles released with 2024-03 has a bad dependency range to
JDT, leading to issues when installing WindowBuilder in old IDEs.

By increasing the minimally required version, an update to WindowBuilder
will also update PDE to a version where this issues has been corrected.

This also updates the version of the imported WindowBuilder features so
that the internal dependency chain remains consistent.