Use FlowLayout instead of GridLayout for property root edit-part
Using a FlowLayout over the GridLayout is the logical choice, given that
each property is supposed to fill a single row. Doing so also avoids the
ugly workaround of explicitly setting the grid-data constraints for each
property figure.

Note that using this layout requires to explicitly enable width and
height tracking in the view port, as otherwise the scrollbars are not
properly shown when the table exceeds the client area of the underlying
figure canvas.