Unify functionality related to the PropertyEditPart
The generic getLevelIndent() method, as well as the property-info
specific methods getTitleX(PropertyInfo), getTitleTextX(PropertyInfo),
and isLocationState(PropertyInfo) methods have been moved from the
PropertyTable to the PropertyEditPart. The latter methods also no longer
expect a PropertyInfo parameter and instead use the edit-part model.

All calls to these methods have been updated to go through the edit
part, as this instance is already known at that point.

The PropertyFigure has been turned into an inner class of the
PropertyEditPart, to emphasize that one class can't exist without the
other and to make it easier for later extraction.