Harmonize property selection with edit-part selection
When a property is selected, this now also selects the corresponding
edit part. Doing so allows us to check whether an edit part is selected
by simply checking the getSelected() flag, rather than having to compare
against the local variable of the property table.

To do so, the setActivePropertyInfo(PropertyInfo) now delegates its call
to select(EditPart), which not only updates the local variable but also
notifies the underlying viewer about this change. This can then be
utilized in the isActiveProperty() method of the PropertyFigure.