Avoid jumping if property table is refreshed by a change
If the table is refreshed due to editing a property, it may happen that
the table is scrolled to the very top and thus hiding the selection.
This is because the active property is set before the new input. Thus
the selection is done on an outdated edit-part.

Instead the input should be set before the selection. Furthermore, the
input needs to be validated, to ensure that the bounds of the figures
are correct.