Update size of TitleFigure whenever the splitter is moved
With to the 4f3e7ed284fa91f6db26f3bdd26eb72f5ff18b89, the property
figures have been split up in a title and value figure. The width of the
title figure is determined by the position of the splitter and therefore
needs to be adjusted, whenever the splitter is moved.