Remove dependency to BurningWave and do reflection via Commons-Lang3
Burningwave is not compatible with Java 24, as it depends on a
Java-internal URLClassPath method whose signature has been changed with
the latest release.

The library was used to resolve the exceptions thrown by CGLib, but is
also required for the call of setAccessible(true) in protected/private
methods, fields and constructors.

Former has become obsolete with the migration to ByteBuddy. Latter has
been resolved by removing most calls to "setAccessible()" and instead
rely on the reflection done by Commons Lang3.

The hack done the ClassLoaderLocalMap, which is used to store
classloader-local properties has been replaced by a weak hash-map that
is kept in the bundle class-loader. Given that weak keys are used, this
should have the same behavior, as the map is disposed together with its
class-loader, which itself is garbage-collected once all other
references have been destroyed.

Closes https://github.com/eclipse-windowbuilder/windowbuilder/issues/1027