Don't use reflection to get the GridBagConstraints
The protected lookupConstraints(Component) method returns a mutable
instance of the GridBagConstraints, as opposed to the copy returned by
the public getConstraints(Component) method.

Instead of accessing the package-private fields, the algorithm for
calculating the cells occupied by each compoent has been extracted from
the GridBagLayout and moved to the GridBagLayoutInfo.

Contributes to
https://github.com/eclipse-windowbuilder/windowbuilder/issues/1027