Remove redundant dispose listener in EmbeddedSwingComposite
The SWT_AWT bridge is explicitly disposed in the dispose() method, so
there is no point to also call it via a dedicated listeners. Depending
on the order in which the listeners are notified, disposing the AWT
frame twice may cause an SWT exception on e.g. Windows.

Note: While this issue has always existed, it only shows up now because
the order in which the frame is disposed has been changed through
9dd68c7a2dc08e664421a05a4d39df5fbfaf1ad6.

Closes https://github.com/eclipse-windowbuilder/windowbuilder/issues/1049