Finalize changes to be compatible with Java 24
This removes the changes done to the ReflectionUtils class that were
made to solve the IllegalAccessException thrown when BurningWave is
disabled (primarily the use of MethodHandles and the FieldUtils class)
and simply tries to set the "accessible" flag via trySetAcccessible().

Furthermore, BurningWave can now be disabled at launch via a system
property, which is enabled by default and explicitly disabled for the
tests.

Contributes to
https://github.com/eclipse-windowbuilder/windowbuilder/issues/1027