Repair EditableSupportTest, NlsSupportTest and SourceEclipseOldTest
The SourceEclipseOldTest was disabled due to a ClassCastException being
thrown. This is because the tests assume that the first source
description of a given JavaInfo is of type EclipseSource. But because
those types are contributed via extension points, the order is
arbitrary.

The NlsSupportTest was disabled because they get stuck on Linux. This
can be avoided by executing them outside the UI thread and by performing
UI access via SWTBot.

One of the tests NlsSupportTest was disabled, even though the test is
fully functional.