Repair and execute SourceCompositeTest
This adapts the SourceCompositeTest to be executed via SWTBot. Note that
for the "test_contextMenu_removeLocale" test, the SourceComposite class
also needs to be adapted.

The current implementation opens a context menu based on the cell on
which the user is clicked. To determine the cell, the current location
of the mouse cursor is used. Because this isn't something that can be
simulated, the location is now stored via the MouseDown event.