Repair and execute the Swing Databinding tests
The databinding tests have been disabled because they rely on the
(withdrawn[1]) JSR295 specification. Given that this specification is
also supported by e.g. Netbeans and the JFormDesigner, it seems too
early to remove it completely.

Given that this artifact is only used for testing, the jar is embedded
directly into the test bundle so that it doesn't need to be added to
Orbit.

With the changes done in 96abbd1ee7ca5f1c04e54a2ea17690badce13da6, an
IllegalArgumentException would be thrown because the array of properties
may contain duplicates. This has been revealed by those tests and fixed
accordingly.

[1] https://jcp.org/en/jsr/detail?id=295