Adapt tests to changed copy-paste functionality
With c7961f3b854a808073e67cf9c56b3307d90bd8d4, the copy-paste operation
remembers the variable name, but the tests were never updated to this
new functionality.

Furthermore, this also fixes a regression where an
IllegalArgumentException would be thrown if the VariableSupport of the
current JavaInfo doesn't override the getName() method.