[Draw2D] Harmonize tool-tip handling of WindowBuilder figures
By wrapping the tool-tip string in a Label, it is possible to use the
setToolTip() method specified by the IFigure interface, thus avoiding an
additional setToolTipText() method.

This Label then needs to be unwrapped in our EventManager, as figures
are not yet supported here, until the manager has been harmonized with
the SWTEventDispatcher.

The tool-tip handling has also been improved so that it is only shown
for non-empty strings.